MSP_HALO_Annotation_Editor 2/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "haloqc"
version = "0.1.0"
description = "HALO tissue separation and bilateral splitting QC tool"
authors = [{name = "Mark St. Pierre"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "PySide6>=6.5",
    "openslide-python>=1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-qt>=4.2",
]

[project.scripts]
haloqc = "haloqc.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["haloqc*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
